[Skipper] Move to LightKernelInterface
