+++ packages/astral/config/config.php
@@ -20,7 +20,6 @@ return static function (ContainerConfigurator $containerConfigurator): void {
 
     $services->load('Symplify\Astral\\', __DIR__ . '/../src')
         ->exclude([
-            __DIR__ . '/../src/HttpKernel',
             __DIR__ . '/../src/StaticFactory',
             __DIR__ . '/../src/ValueObject',
             __DIR__ . '/../src/NodeVisitor',
