--- packages/astral/src/HttpKernel/AstralKernel.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\Astral\HttpKernel;
-
-use Symfony\Component\Config\Loader\LoaderInterface;
-use Symplify\SymplifyKernel\HttpKernel\AbstractSymplifyKernel;
-
-final class AstralKernel extends AbstractSymplifyKernel
-{
-    public function registerContainerConfiguration(LoaderInterface $loader): void
-    {
-        $loader->load(__DIR__ . '/../../config/config.php');
-    }
-}
