+++ packages/astral/src/ValueObject/AstralConfig.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\Astral\ValueObject;
+
+final class AstralConfig
+{
+    /**
+     * @var string
+     */
+    public const FILE_PATH = __DIR__ . '/../../config/config.php';
+}
