+++ packages/astral/tests/HttpKernel/AstralKernel.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Symplify\MarkdownDiff\Tests\HttpKernel;
+namespace Symplify\Astral\Tests\HttpKernel;
 
 use Psr\Container\ContainerInterface;
-use Symplify\MarkdownDiff\ValueObject\MarkdownDiffConfig;
+use Symplify\Astral\ValueObject\AstralConfig;
 use Symplify\SymplifyKernel\HttpKernel\AbstractSymplifyKernel;
 
-final class MarkdownDiffKernel extends AbstractSymplifyKernel
+final class AstralKernel extends AbstractSymplifyKernel
 {
     public function createFromConfigs(array $configFiles): ContainerInterface
     {
-        $configFiles[] = MarkdownDiffConfig::FILE_PATH;
+        $configFiles[] = AstralConfig::FILE_PATH;
         return $this->create([], [], $configFiles);
     }
 }
