+++ packages/easy-ci/composer.json
@@ -10,7 +10,6 @@
         "composer/semver": "^3.2",
         "symfony/dependency-injection": "^5.3|^6.0",
         "symfony/console": "^5.3|^6.0",
-        "symfony/http-kernel": "^5.3|^6.0",
         "nette/utils": "^3.2",
         "nikic/php-parser": "^4.13",
         "symplify/composer-json-manipulator": "^9.5",
