+++ packages/markdown-diff/src/ValueObject/MarkdownDiffConfig.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\MarkdownDiff\ValueObject;
+
+final class MarkdownDiffConfig
+{
+    /**
+     * @var string
+     */
+    public const FILE_PATH = __DIR__ . '/../../config/config.php';
+}
