+++ packages/skipper/src/ValueObject/SkipperConfig.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\Skipper\ValueObject;
+
+final class SkipperConfig
+{
+    /**
+     * @var string
+     */
+    public const FILE_PATH = __DIR__ . '/../../config/config.php';
+}
