+++ packages/skipper/tests/HttpKernel/SkipperKernel.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Symplify\MarkdownDiff\Tests\HttpKernel;
+namespace Symplify\Skipper\Tests\HttpKernel;
 
 use Psr\Container\ContainerInterface;
-use Symplify\MarkdownDiff\ValueObject\MarkdownDiffConfig;
+use Symplify\Skipper\ValueObject\SkipperConfig;
 use Symplify\SymplifyKernel\HttpKernel\AbstractSymplifyKernel;
 
-final class MarkdownDiffKernel extends AbstractSymplifyKernel
+final class SkipperKernel extends AbstractSymplifyKernel
 {
     public function createFromConfigs(array $configFiles): ContainerInterface
     {
-        $configFiles[] = MarkdownDiffConfig::FILE_PATH;
+        $configFiles[] = SkipperConfig::FILE_PATH;
         return $this->create([], [], $configFiles);
     }
 }
