+++ packages/skipper/tests/Skipper/Only/OnlySkipperTest.php
@@ -6,8 +6,8 @@ namespace Symplify\Skipper\Tests\Skipper\Only;
 
 use Iterator;
 use Symplify\PackageBuilder\Testing\AbstractKernelTestCase;
-use Symplify\Skipper\HttpKernel\SkipperKernel;
 use Symplify\Skipper\Skipper\Skipper;
+use Symplify\Skipper\Tests\HttpKernel\SkipperKernel;
 use Symplify\Skipper\Tests\Skipper\Only\Source\IncludeThisClass;
 use Symplify\Skipper\Tests\Skipper\Only\Source\SkipCompletely;
 use Symplify\Skipper\Tests\Skipper\Only\Source\SkipCompletelyToo;
