+++ packages/skipper/tests/Skipper/Skip/SkipSkipperTest.php
@@ -6,8 +6,8 @@ namespace Symplify\Skipper\Tests\Skipper\Skip;
 
 use Iterator;
 use Symplify\PackageBuilder\Testing\AbstractKernelTestCase;
-use Symplify\Skipper\HttpKernel\SkipperKernel;
 use Symplify\Skipper\Skipper\Skipper;
+use Symplify\Skipper\Tests\HttpKernel\SkipperKernel;
 use Symplify\Skipper\Tests\Skipper\Skip\Source\AnotherClassToSkip;
 use Symplify\Skipper\Tests\Skipper\Skip\Source\NotSkippedClass;
 use Symplify\Skipper\Tests\Skipper\Skip\Source\SomeClassToSkip;
